refactor: Move approval utils to fixtures
